move files from system that don't exist in skeleton
