+++ skeleton/fix.c
