+++ skeleton/kbd.c
