+++ skeleton/kbd.h
