+++ skeleton/libfixmath/fixmath.h
@@ -1,0 +1,23 @@
+#ifndef __libfixmath_fixmath_h__
+#define __libfixmath_fixmath_h__
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif
+
+/*!
+	\file fixmath.h
+	\brief Functions to perform fast accurate fixed-point math operations.
+*/
+
+#include "uint32.h"
+#include "int64.h"
+#include "fract32.h"
+#include "fix16.h"
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
