+++ skeleton/libfixmath/license.txt
