+++ skeleton/screen.c
