--- system/fonts/dejavu_numerals_24.c
@@ -1,112 +1,0 @@
-/* output from xpm-convert.c */ 
-/* first two bytes are width and height */ 
-
-/* data follows */ 
-#include "dejavu_numerals_24.h"
-
-#if 1
-
-
-
-// width (16) * height (16) plus descriptors
-//#define DEJAVU_NUMERALS_LEN
-
-//const char dejavu_numerals_24[][] = { 
-
-/* const unsigned char DejaVuSansMono_Bold_18_24_30_0_xpm_glyph[] =  */
-
-
-const dejavu_24_glyph font_dejavu_24_numerals[11] = { 
-
-  { .raw = {
-      18 , 24, 
-0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x06 , 0x03 , 0x00 , 0x03 , 0x06 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x05 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x05 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x05 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x05 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x09 , 0x00 , 0x00 , 0x00 , 0x07 , 0x0C , 0x0E , 0x0C , 0x07 , 0x00 , 0x00 , 0x00 , 0x09 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x01 , 0x00 , 0x00 , 0x08 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x08 , 0x00 , 0x00 , 0x02 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0B , 0x00 , 0x00 , 0x03 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x03 , 0x00 , 0x00 , 0x0B , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x06 , 0x00 , 0x00 , 0x0B , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x00 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x04 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x00 , 0x00 , 0x04 , 0x0F , 0x0F , 0x0F , 0x01 , 0x00 , 0x00 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x01 , 0x00 , 0x00 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x04 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x00 , 0x00 , 0x04 , 0x0F , 0x0F , 0x0F , 0x06 , 0x00 , 0x00 , 0x0B , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x00 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x0B , 0x00 , 0x00 , 0x03 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x03 , 0x00 , 0x00 , 0x0B , 0x0F , 0x0F , 0x0F , 0x0E , 0x01 , 0x00 , 0x00 , 0x08 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x08 , 0x00 , 0x00 , 0x02 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x09 , 0x00 , 0x00 , 0x00 , 0x07 , 0x0C , 0x0E , 0x0C , 0x07 , 0x00 , 0x00 , 0x00 , 0x09 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x05 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x05 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x05 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x05 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x06 , 0x02 , 0x00 , 0x02 , 0x06 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 
-
-    }  } ,
-
-
-  { .raw = {
-      18 , 24,
-      0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0A , 0x08 , 0x06 , 0x04 , 0x02 , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x01 , 0x03 , 0x05 , 0x07 , 0x09 , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 0x0B , 
-
-    }  } ,
-
-
-  { .raw = {
-      18 , 24,
-      0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x09 , 0x06 , 0x04 , 0x02 , 0x01 , 0x00 , 0x01 , 0x04 , 0x06 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x04 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x02 , 0x05 , 0x08 , 0x0B , 0x0D , 0x0F , 0x0F , 0x0E , 0x09 , 0x03 , 0x00 , 0x00 , 0x00 , 0x08 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x04 , 0x00 , 0x00 , 0x04 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x00 , 0x00 , 0x02 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x00 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x02 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x03 , 0x00 , 0x00 , 0x03 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x02 , 0x00 , 0x00 , 0x02 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x02 , 0x00 , 0x00 , 0x02 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x01 , 0x00 , 0x00 , 0x02 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x01 , 0x00 , 0x00 , 0x02 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x00 , 0x02 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x09 , 0x00 , 0x00 , 0x00 , 0x02 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x08 , 0x00 , 0x00 , 0x00 , 0x03 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0D , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0D , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 
-
-    }  } ,
-
-
-  { .raw = {
-      18 , 24,
-      0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x08 , 0x05 , 0x02 , 0x00 , 0x00 , 0x01 , 0x03 , 0x08 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x01 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x02 , 0x06 , 0x09 , 0x0C , 0x0E , 0x0E , 0x0B , 0x06 , 0x00 , 0x00 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x08 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x0B , 0x07 , 0x00 , 0x00 , 0x01 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x03 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x04 , 0x0B , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x04 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x09 , 0x03 , 0x00 , 0x00 , 0x02 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x05 , 0x00 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x04 , 0x00 , 0x00 , 0x04 , 0x0F , 0x0F , 0x0F , 0x0F , 0x02 , 0x07 , 0x0A , 0x0C , 0x0E , 0x0F , 0x0E , 0x0B , 0x08 , 0x02 , 0x00 , 0x00 , 0x00 , 0x09 , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x08 , 0x05 , 0x02 , 0x00 , 0x00 , 0x00 , 0x01 , 0x04 , 0x07 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 
-
-    }  }, 
-
-
-  { .raw = {
-      18 , 24,
-      0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x01 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x04 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x09 , 0x00 , 0x01 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x01 , 0x00 , 0x09 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x05 , 0x00 , 0x04 , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x09 , 0x00 , 0x01 , 0x0C , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x01 , 0x00 , 0x08 , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x06 , 0x00 , 0x03 , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x0B , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x02 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x06 , 0x00 , 0x02 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x02 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x01 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 
-
-    }  }, 
-
-
-  { .raw = {
-      18 , 24,
-      0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x06 , 0x02 , 0x01 , 0x01 , 0x04 , 0x07 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x08 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x03 , 0x08 , 0x0B , 0x0E , 0x0F , 0x0E , 0x0B , 0x07 , 0x01 , 0x00 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x02 , 0x00 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x03 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x03 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x02 , 0x00 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x0F , 0x03 , 0x07 , 0x0A , 0x0D , 0x0E , 0x0F , 0x0E , 0x0B , 0x07 , 0x01 , 0x00 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x08 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x01 , 0x09 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x08 , 0x05 , 0x03 , 0x01 , 0x00 , 0x01 , 0x02 , 0x05 , 0x08 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 
-
-    }  }, 
-
-
-  { .raw = {
-      18 , 24,
-      0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x05 , 0x02 , 0x01 , 0x00 , 0x01 , 0x03 , 0x07 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x04 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x02 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x03 , 0x00 , 0x00 , 0x00 , 0x04 , 0x0A , 0x0D , 0x0F , 0x0D , 0x0B , 0x06 , 0x02 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x00 , 0x00 , 0x09 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x01 , 0x00 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x09 , 0x00 , 0x00 , 0x00 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x05 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x03 , 0x00 , 0x00 , 0x08 , 0x0F , 0x09 , 0x03 , 0x01 , 0x00 , 0x02 , 0x05 , 0x0B , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x01 , 0x00 , 0x00 , 0x0C , 0x05 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x06 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x02 , 0x09 , 0x0D , 0x0F , 0x0D , 0x09 , 0x02 , 0x00 , 0x00 , 0x00 , 0x0B , 0x0F , 0x0F , 0x0F , 0x01 , 0x00 , 0x00 , 0x03 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x03 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x01 , 0x00 , 0x00 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x02 , 0x0F , 0x0F , 0x0F , 0x04 , 0x00 , 0x00 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x06 , 0x00 , 0x00 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x02 , 0x0F , 0x0F , 0x0F , 0x0F , 0x01 , 0x00 , 0x03 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x03 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x08 , 0x00 , 0x00 , 0x02 , 0x09 , 0x0D , 0x0F , 0x0D , 0x09 , 0x02 , 0x00 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x04 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x04 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x04 , 0x01 , 0x00 , 0x00 , 0x02 , 0x06 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 
-
-    }  }, 
-
-
-  { .raw = {
-      18 , 24,
-      0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x03 , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x08 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x01 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x05 , 0x00 , 0x00 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x00 , 0x00 , 0x01 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x08 , 0x00 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x02 , 0x00 , 0x00 , 0x0B , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x00 , 0x00 , 0x03 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x06 , 0x00 , 0x00 , 0x08 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x01 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x09 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x04 , 0x00 , 0x00 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x00 , 0x00 , 0x01 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x02 , 0x00 , 0x00 , 0x0B , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x00 , 0x00 , 0x02 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x06 , 0x00 , 0x00 , 0x08 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x01 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x09 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x04 , 0x00 , 0x00 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 
-
-    }  }, 
-
-
-  { .raw = {
-      18 , 24,
-      0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x07 , 0x04 , 0x01 , 0x00 , 0x01 , 0x04 , 0x07 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x01 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x01 , 0x0B , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x06 , 0x00 , 0x00 , 0x00 , 0x07 , 0x0D , 0x0F , 0x0D , 0x07 , 0x00 , 0x00 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x02 , 0x00 , 0x00 , 0x09 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x08 , 0x00 , 0x00 , 0x02 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x02 , 0x00 , 0x00 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x02 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x06 , 0x00 , 0x00 , 0x09 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x08 , 0x00 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x01 , 0x00 , 0x01 , 0x07 , 0x0D , 0x0F , 0x0D , 0x07 , 0x00 , 0x00 , 0x02 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0D , 0x04 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x04 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x05 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x05 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x05 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x05 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x03 , 0x00 , 0x00 , 0x04 , 0x0A , 0x0D , 0x0F , 0x0D , 0x0A , 0x04 , 0x00 , 0x00 , 0x03 , 0x0E , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x05 , 0x00 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x03 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x00 , 0x00 , 0x03 , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x01 , 0x00 , 0x00 , 0x0C , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x04 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x05 , 0x00 , 0x00 , 0x04 , 0x0F , 0x0F , 0x0F , 0x08 , 0x00 , 0x00 , 0x00 , 0x04 , 0x0A , 0x0D , 0x0F , 0x0D , 0x0A , 0x04 , 0x00 , 0x00 , 0x00 , 0x08 , 0x0F , 0x0F , 0x0F , 0x0F , 0x03 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x03 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x04 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x04 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0B , 0x06 , 0x03 , 0x01 , 0x00 , 0x01 , 0x03 , 0x06 , 0x0B , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 
-
-    }  }, 
-
-
-  { .raw = {
-      18 , 24,
-      0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x06 , 0x02 , 0x00 , 0x00 , 0x01 , 0x04 , 0x09 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x04 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x07 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x04 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x00 , 0x00 , 0x00 , 0x02 , 0x08 , 0x0D , 0x0F , 0x0D , 0x08 , 0x02 , 0x00 , 0x00 , 0x08 , 0x0F , 0x0F , 0x0F , 0x0F , 0x05 , 0x00 , 0x00 , 0x03 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x03 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x0F , 0x02 , 0x00 , 0x00 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x0A , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x06 , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x04 , 0x0F , 0x0F , 0x0F , 0x01 , 0x00 , 0x00 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x02 , 0x0F , 0x0F , 0x0F , 0x05 , 0x00 , 0x00 , 0x03 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x03 , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x0A , 0x00 , 0x00 , 0x00 , 0x02 , 0x09 , 0x0D , 0x0F , 0x0D , 0x09 , 0x02 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x05 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x06 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x05 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x05 , 0x0B , 0x00 , 0x00 , 0x01 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0A , 0x05 , 0x01 , 0x00 , 0x00 , 0x03 , 0x09 , 0x0F , 0x08 , 0x00 , 0x00 , 0x03 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x05 , 0x00 , 0x00 , 0x05 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0E , 0x00 , 0x00 , 0x00 , 0x08 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x06 , 0x00 , 0x00 , 0x01 , 0x0E , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x08 , 0x00 , 0x00 , 0x00 , 0x07 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x02 , 0x07 , 0x0B , 0x0E , 0x0F , 0x0D , 0x0A , 0x04 , 0x00 , 0x00 , 0x00 , 0x03 , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x02 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x04 , 0x0D , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0C , 0x07 , 0x03 , 0x01 , 0x00 , 0x00 , 0x02 , 0x05 , 0x0A , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 0x0F , 
-
-    }  }, 
-
-
-  {     .raw = {
-      18 , 24,
-      0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x0F , 0x0F , 0x0F , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x00 , 0x01 , 
-
-    }  }
-};
-
-/* const unsigned char ** font_dejavu_numerals_24 = { */
-/*   DejaVuSansMono_Bold_18_24_30_dot_xpm_glyph, */
-/*   DejaVuSansMono_Bold_18_24_30_0_xpm_glyph, */
-/*   DejaVuSansMono_Bold_18_24_30_1_xpm_glyph, */
-/*   DejaVuSansMono_Bold_18_24_30_2_xpm_glyph, */
-/*   DejaVuSansMono_Bold_18_24_30_3_xpm_glyph, */
-/*   DejaVuSansMono_Bold_18_24_30_4_xpm_glyph, */
-/*   DejaVuSansMono_Bold_18_24_30_5_xpm_glyph, */
-/*   DejaVuSansMono_Bold_18_24_30_6_xpm_glyph, */
-/*   DejaVuSansMono_Bold_18_24_30_7_xpm_glyph, */
-/*   DejaVuSansMono_Bold_18_24_30_8_xpm_glyph, */
-/*   DejaVuSansMono_Bold_18_24_30_9_xpm_glyph, */
-/* }; */
-
-#endif
