--- system/libfixmath/uint32.h
@@ -1,19 +1,0 @@
-#ifndef __libfixmath_uint32_h__
-#define __libfixmath_uint32_h__
-
-#ifdef __cplusplus
-extern "C"
-{
-#endif
-
-#include <stdint.h>
-
-/*! Performs an unsigned log-base2 on the specified unsigned integer and returns the result.
-*/
-extern uint32_t uint32_log2(uint32_t inVal);
-
-#ifdef __cplusplus
-}
-#endif
-
-#endif
